separated switch dialog from window file
